Use transformation stack for draw() and update()
Before drawing and updating we get the chance to set any
transformations (translate, rotate, scale shear) and they
will set the geometrical context for everthing. All touches
will be transformed to this coordinate space to be converted
to grid position. This could be used for fake 3D.

The goal is that only 'main' module should know about screen
info like resolution and DPI, the rest of code is 'normalized'.