Drop support for touch pressure
Touch pressure detection works only on few phones and not reliably
enough to be used for controling tone. It is removed from code.

There are still two instruments using per-note velocity, which is
controlled by tilting the phone. In case of 'scat', it is not even
velocity but completely different sample set abusing velocity mechanics.